requirements.txt
mss>=0.9.0
pydirectinput>=1.0.4
pystray>=0.19.0
Pillow>=10.0.0
keyboard>=0.13.5
